Add documentation and basic examples to improve the file structure
