+++ demo_manufacturer/demo_model/requirements.txt
@@ -1,0 +1,1 @@
+mcp==1.9.4
